move logic to index file
